Feat: add analyzable module
